feat: 선택 상품 레포지토리 추가

+++ src/main/java/com/virtukch/dongiveupbe/select_product/repository/SelectProductRepository.java
@@ -3,5 +3,5 @@ package com.virtukch.dongiveupbe.select_product.repository;
 import com.virtukch.dongiveupbe.select_product.entity.SelectProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SelectProductRepository extends JpaRepository<SelectProduct, Integer> {
+public interface SelectProductRepository extends JpaRepository<SelectProduct, Long> {
 }
